Restructured PDFjoiner
